Fixed "Delete Account" Functionality
Because the user is now linked to so many tables via foreign key, I had to also remove all of those entries from the other tables in order to be able to delete the user (i.e. delete all attendance indication entries for the user, all events created by the user, and all images related to the events created by the user).

Very poor optimization. Will optimize further in the future.